refactor: 使用單數 user
