+++ src/auth/auth.module.ts
@@ -1,13 +1,13 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { UserEntity } from "src/users/entities/user.entity";
-import { UsersModule } from "src/users/users.module";
+import { UserEntity } from "src/user/entities/user.entity";
+import { UserModule } from "src/user/user.module";
 
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
 
 @Module({
-  imports: [UsersModule, TypeOrmModule.forFeature([UserEntity])],
+  imports: [UserModule, TypeOrmModule.forFeature([UserEntity])],
   controllers: [AuthController],
   providers: [AuthService],
 })
