+++ src/user/user.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+
+import { UserService } from "./user.service";
+
+@Module({
+  providers: [UserService],
+  exports: [UserService],
+})
+export class UserModule {}
